+++ ext/twisty_puzzles/native/cube_state.h
@@ -5,7 +5,7 @@
 #include "face_symbols.h"
 
 typedef struct {
-  size_t cube_size;
+  long cube_size;
   VALUE* stickers;
 } CubeStateData;
 
@@ -24,7 +24,7 @@ extern const rb_data_type_t CubeStateData_type;
     } \
   } while(0)
 
-void rotate_slice_for_cube(face_index_t turned_face_index, size_t slice_index, direction_t direction, const CubeStateData* data);
+void rotate_slice_for_cube(face_index_t turned_face_index, long slice_index, direction_t direction, const CubeStateData* data);
 
 void rotate_face_for_cube(face_index_t turned_face_index, direction_t direction, const CubeStateData* data);
